Wasapi to wasapi RTP works now. Commented out a bit of wexception because of some access violation.

+++ wasapi/AudioConverter.cpp
@@ -185,7 +185,7 @@ HRESULT AudioConverter::Transform(IMediaSample* pIn, IMediaSample* pOut)
 		}
 	}
 
-	pOut->SetActualDataLength(pbWrite - pbOut);
+	pOut->SetActualDataLength((long)(pbWrite - pbOut));
 
 	return S_OK;
 }
